Remove memory tracking code
