+++ code/globalincs/fsmemory.h
@@ -2,12 +2,6 @@
 #ifndef _FSMEMORY_H
 #define _FSMEMORY_H
 
-#ifndef NDEBUG
-	#include "globalincs/memory/debug.h"
-#else
-	#include "globalincs/memory/release.h"
-#endif
-
 #include "globalincs/memory/memory.h"
 #include "globalincs/memory/utils.h"
