+++ code/globalincs/memory/memory.h
@@ -2,21 +2,47 @@
 #pragma once
 
 #include <stddef.h>
+#include <cstdlib>
+
+#include "globalincs/pstypes.h"
 
 namespace memory
 {
-	void init();
-
-	size_t get_used_memory();
-
 	struct quiet_alloc_t { quiet_alloc_t(){} };
 	extern const quiet_alloc_t quiet_alloc;
 
-	void register_malloc(size_t size, const char *filename, int line);
-	void unregister_malloc(size_t size, const char *filename, int line);
+	void out_of_memory();
+}
 
-	size_t sort_memory_blocks();
-	bool get_memory_block(size_t index, const char*& fileName, size_t& usedMemory);
+inline void *vm_malloc(size_t size, const memory::quiet_alloc_t &)
+{ return std::malloc(size); }
 
-	void out_of_memory(const char* filename, int line);
+inline void *vm_malloc(size_t size)
+{
+	auto ptr = vm_malloc(size, memory::quiet_alloc);
+
+	if (ptr == NULL)
+	{
+		memory::out_of_memory();
+	}
+
+	return ptr;
+}
+
+inline void vm_free(void *ptr)
+{ std::free(ptr); }
+
+inline void *vm_realloc(void *ptr, size_t size, const memory::quiet_alloc_t &)
+{ return std::realloc(ptr, size); }
+
+inline void *vm_realloc(void *ptr, size_t size)
+{
+	auto ret_ptr = vm_realloc(ptr, size, memory::quiet_alloc);
+
+	if (ret_ptr == NULL)
+	{
+		memory::out_of_memory();
+	}
+
+	return ret_ptr;
 }
