+++ freespace2/freespace.cpp
@@ -260,10 +260,8 @@ fix FrametimeOverall = 0;
 
 #ifndef NDEBUG
 	int	Show_framerate = 1;
-	int	Show_mem = 1;
 #else 
 	int	Show_framerate = 0;
-	int	Show_mem = 0;
 #endif
 
 int	Framerate_cap = 120;
@@ -1191,42 +1189,6 @@ void game_loading_callback(int count)
 	}
 #endif
 
-#ifndef NDEBUG
-	if(Cmdline_show_mem_usage)
-	{
-		int line_height = gr_get_font_height() + 1;
-		size_t i;
-		char mem_buffer[1000];
-
-		memory::sort_memory_blocks();
-		for(i = 0; i < 30; i++)
-		{
-			const char* full_name;
-			size_t size;
-			if (!memory::get_memory_block(i, full_name, size))
-			{
-				break;
-			}
-
-			size /= 1024;
-
-			if(size == 0)
-				break;
-
-			const char *short_name = strrchr(full_name, DIR_SEPARATOR_CHAR);
-			if(short_name == NULL)
-				short_name = full_name;
-			else
-				short_name++;
-
-			sprintf(mem_buffer,"%s:\t" SIZE_T_ARG " K", short_name, size);
-			gr_string( 20, 220 + (int)(i*line_height), mem_buffer, GR_RESIZE_MENU);
-		}
-		sprintf(mem_buffer,"Total RAM:\t" SIZE_T_ARG " K", memory::get_used_memory() / 1024);
-		gr_string( 20, 230 + (int)(i*line_height), mem_buffer, GR_RESIZE_MENU);
-	}
-#endif	// !NDEBUG
-
 	os_ignore_events();
 
 	if (do_flip)
@@ -1547,37 +1509,6 @@ DCF(warp, "Tests warpin effect")
 	
 }
 
-DCF(show_mem,"Toggles showing mem usage")
-{
-	bool process = true;
-
-	if (dc_optional_string_either("help", "--help")) {
-		dc_printf( "Usage: (optional) bool Show_mem\n If true, Show_mem is set and Show_cpu is cleared.  If false, then Show_mem is cleared.  If nothing passed, then toggle.\n" );
-		process = false;
-	}
-	
-	if (dc_optional_string_either("status", "--status") || dc_optional_string_either("?", "--?")) {
-		dc_printf("Show_mem is %s\n", (Show_mem ? "TRUE" : "FALSE"));
-		dc_printf("Show_cpu is %s\n", (Show_cpu ? "TRUE" : "FALSE"));
-		process = false;
-	}
-	
-	if (!process) {
-		// Help and/or status was given, so don't process the command
-		return;
-	} // Else, process the command
-
-	if (!dc_maybe_stuff_boolean(&Show_mem)) {
-		// Nothing passed, so toggle
-		Show_mem = !Show_mem;
-	}	// Else, value was set/cleared by user
-
-	// Can't show mem and cpu at same time
-	if (Show_mem) {
-		Show_cpu = false;
-	}
-}
-
 DCF(show_cpu,"Toggles showing cpu usage")
 {
 	bool process = true;
@@ -1589,7 +1520,6 @@ DCF(show_cpu,"Toggles showing cpu usage")
 	
 	if (dc_optional_string_either("status", "--status") || dc_optional_string_either("?", "--?")) {
 		dc_printf("Show_cpu is %s\n", (Show_cpu ? "TRUE" : "FALSE"));
-		dc_printf("Show_mem is %s\n", (Show_mem ? "TRUE" : "FALSE"));
 		process = false;
 	}
 
@@ -1602,11 +1532,6 @@ DCF(show_cpu,"Toggles showing cpu usage")
 		// Nothing passed, so toggle
 		Show_cpu = !Show_cpu;
 	}	// Else, value was set/cleared by user
-
-	// Can't show mem and cpu at same time
-	if (Show_cpu) {
-		Show_mem = false;
-	}
 }
 
 #endif
@@ -2275,80 +2200,12 @@ void game_show_framerate()
 			sy += line_height;
 		}
 	}
-	 	
-	if ( Show_mem  ) {
-
-		int sx,sy;
-		sx = gr_screen.center_offset_x + gr_screen.center_w - 154;
-		sy = gr_screen.center_offset_y + 15;
-
-		gr_set_color_fast(&HUD_color_debug);
-
-		{
-			gr_printf_no_resize( sx, sy, NOX("DYN: " SIZE_T_ARG " KB\n"), memory::get_used_memory()/1024 );
-			sy += line_height;
-		}	
-
-		{
-			extern int Model_ram;
-			gr_printf_no_resize( sx, sy, NOX("POF: %d KB\n"), Model_ram/1024 );
-			sy += line_height;
-		}	
-
-		gr_printf_no_resize( sx, sy, NOX("%s: %d KB\n"), (Cmdline_cache_bitmaps) ? NOX("C-BMP") : NOX("BMP"), bm_texture_ram/1024 );
-		sy += line_height;
-
-		gr_printf_no_resize( sx, sy, NOX("S-SRAM: %d KB\n"), Snd_sram/1024 );		// mem used to store game sound
-		sy += line_height;
-
-		{
-			extern int GL_textures_in;
-			extern size_t GL_vertex_data_in;
-			gr_printf_no_resize( sx, sy, NOX("VRAM: " SIZE_T_ARG " KB\n"), (GL_textures_in + GL_vertex_data_in)/1024 );
-			sy += line_height;
-		}
-	}
-
 
 	if ( Show_player_pos ) {
 		int sx, sy;
 		sx = gr_screen.center_offset_x + 320;
 		sy = gr_screen.center_offset_y + 100;
 		gr_printf_no_resize(sx, sy, NOX("Player Pos: (%d,%d,%d)"), fl2i(Player_obj->pos.xyz.x), fl2i(Player_obj->pos.xyz.y), fl2i(Player_obj->pos.xyz.z));
-	}
-
-
-	if(Cmdline_show_mem_usage)
-	{
-		size_t i;
-		char mem_buffer[1000];
-
-		memory::sort_memory_blocks();
-		for(i = 0; i < 30; i++)
-		{
-			const char* full_name;
-			size_t size;
-			if (!memory::get_memory_block(i, full_name, size))
-			{
-				break;
-			}
-
-			size /= 1024;
-
-			if(size == 0)
-				break;
-
-			const char *short_name = strrchr(full_name, DIR_SEPARATOR_CHAR);
-			if(short_name == NULL)
-				short_name = full_name;
-			else
-				short_name++;
-
-			sprintf(mem_buffer,"%s:\t" SIZE_T_ARG " K", short_name, size);
-			gr_string( 20, 220 + (int)(i*line_height), mem_buffer, GR_RESIZE_MENU);
-		}
-		sprintf(mem_buffer,"Total RAM:\t" SIZE_T_ARG " K", memory::get_used_memory() / 1024);
-		gr_string( 20, 230 + (int)(i*line_height), mem_buffer, GR_RESIZE_MENU);
 	}
 
 	MONITOR_INC(NumPolys, modelstats_num_polys);
@@ -8583,8 +8440,6 @@ int actual_main(int argc, char *argv[])
 {
 	int result = -1;
 	Assert(argc > 0);
-
-	memory::init();
 
 #ifdef WIN32
 	// Don't let more than one instance of FreeSpace run.
